requirements.txt
fastapi
uvicorn
pytorch
torchvision
albumentations
python-multipart